+++ CHANGELOG.md
@@ -6,7 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- [Bugsnag] Added new tool to get the project collaborators.
+- [Bugsnag] Added new tool to list project collaborators.
 
 ### Added